2.6.0 release

+++ tests/helics/application_api/LoggingTests.cpp
@@ -8,7 +8,6 @@ SPDX-License-Identifier: BSD-3-Clause
 #include "helics/application_api/ValueFederate.hpp"
 #include "helics/core/BrokerFactory.hpp"
 //#include "helics/core/CoreFactory.hpp"
-#include "helics/common/logger.h"
 #include "helics/core/Core.hpp"
 #include "helics/core/core-exceptions.hpp"
 #include "helics/core/helics_definitions.hpp"
@@ -32,9 +31,10 @@ TEST(logging_tests, basic_logging)
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
 
     gmlc::libguarded::guarded<std::vector<std::pair<int, std::string>>> mlog;
-    Fed->setLoggingCallback([&mlog](int level, const std::string&, const std::string& message) {
-        mlog.lock()->emplace_back(level, message);
-    });
+    Fed->setLoggingCallback(
+        [&mlog](int level, const std::string& /*unused*/, const std::string& message) {
+            mlog.lock()->emplace_back(level, message);
+        });
     Fed->logMessage(3, "test log message");
     Fed->enterExecutingMode();
     Fed->finalize();
@@ -44,11 +44,26 @@ TEST(logging_tests, basic_logging)
 
 TEST(logging_tests, file_logging)
 {
+    const std::string lfilename = "logfile.txt";
+    if (ghc::filesystem::exists(lfilename)) {
+        std::error_code ec;
+        bool res = ghc::filesystem::remove(lfilename, ec);
+        int ii = 0;
+        while (!res) {
+            std::this_thread::sleep_for(std::chrono::milliseconds(100));
+            res = ghc::filesystem::remove(lfilename, ec);
+            ++ii;
+            if (ii > 15) {
+                break;
+            }
+        }
+        EXPECT_TRUE(res);
+    }
     helics::FederateInfo fi(CORE_TYPE_TO_TEST);
     fi.coreInitString = "--autobroker --logfile logfile.txt --fileloglevel=5";
 
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
-    const std::string lfilename = "logfile.txt";
+
     Fed->enterExecutingMode();
     Fed->finalize();
     auto cr = Fed->getCorePointer();
@@ -59,27 +74,32 @@ TEST(logging_tests, file_logging)
     cr.reset();
     helics::cleanupHelicsLibrary();
     std::error_code ec;
-    bool res = ghc::filesystem::remove(lfilename, ec);
-    int ii = 0;
-    while (!res) {
-        std::this_thread::sleep_for(std::chrono::milliseconds(100));
-        res = ghc::filesystem::remove(lfilename, ec);
-        ++ii;
-        if (ii > 15) {
-            break;
-        }
-    }
-    EXPECT_TRUE(res);
+    ghc::filesystem::remove(lfilename, ec);
 }
 
 TEST(logging_tests, file_logging_p2)
 {
+    const std::string lfilename = "logfile2.txt";
+    if (ghc::filesystem::exists(lfilename)) {
+        std::error_code ec;
+        bool res = ghc::filesystem::remove(lfilename, ec);
+        int ii = 0;
+        while (!res) {
+            std::this_thread::sleep_for(std::chrono::milliseconds(100));
+            res = ghc::filesystem::remove(lfilename, ec);
+            ++ii;
+            if (ii > 15) {
+                break;
+            }
+        }
+        EXPECT_TRUE(res);
+    }
     helics::FederateInfo fi(CORE_TYPE_TO_TEST);
     fi.coreInitString = "--autobroker --fileloglevel=5";
 
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
     auto cr = Fed->getCorePointer();
-    const std::string lfilename = "logfile2.txt";
+
     cr->setLogFile(lfilename);
     Fed->enterExecutingMode();
     Fed->finalize();
@@ -90,17 +110,7 @@ TEST(logging_tests, file_logging_p2)
     cr.reset();
     helics::cleanupHelicsLibrary();
     std::error_code ec;
-    bool res = ghc::filesystem::remove(lfilename, ec);
-    int ii = 0;
-    while (!res) {
-        std::this_thread::sleep_for(std::chrono::milliseconds(100));
-        res = ghc::filesystem::remove(lfilename, ec);
-        ++ii;
-        if (ii > 15) {
-            break;
-        }
-    }
-    EXPECT_TRUE(res);
+    ghc::filesystem::remove(lfilename, ec);
 }
 
 TEST(logging_tests, check_log_message)
@@ -112,9 +122,10 @@ TEST(logging_tests, check_log_message)
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
 
     gmlc::libguarded::guarded<std::vector<std::pair<int, std::string>>> mlog;
-    Fed->setLoggingCallback([&mlog](int level, const std::string&, const std::string& message) {
-        mlog.lock()->emplace_back(level, message);
-    });
+    Fed->setLoggingCallback(
+        [&mlog](int level, const std::string& /*unused*/, const std::string& message) {
+            mlog.lock()->emplace_back(level, message);
+        });
 
     Fed->enterExecutingMode();
     Fed->logInfoMessage("test MEXAGE");
@@ -140,9 +151,10 @@ TEST(logging_tests, check_log_message_functions)
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
 
     gmlc::libguarded::guarded<std::vector<std::pair<int, std::string>>> mlog;
-    Fed->setLoggingCallback([&mlog](int level, const std::string&, const std::string& message) {
-        mlog.lock()->emplace_back(level, message);
-    });
+    Fed->setLoggingCallback(
+        [&mlog](int level, const std::string& /*unused*/, const std::string& message) {
+            mlog.lock()->emplace_back(level, message);
+        });
 
     Fed->enterExecutingMode();
     Fed->logErrorMessage("test ERROR");
@@ -187,9 +199,10 @@ TEST(logging_tests, check_log_message_levels)
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
 
     gmlc::libguarded::guarded<std::vector<std::pair<int, std::string>>> mlog;
-    Fed->setLoggingCallback([&mlog](int level, const std::string&, const std::string& message) {
-        mlog.lock()->emplace_back(level, message);
-    });
+    Fed->setLoggingCallback(
+        [&mlog](int level, const std::string& /*unused*/, const std::string& message) {
+            mlog.lock()->emplace_back(level, message);
+        });
 
     Fed->enterExecutingMode();
     Fed->logMessage(3, "test MEXAGE1");
@@ -220,9 +233,10 @@ TEST(logging_tests, check_log_message_levels_high)
     auto Fed = std::make_shared<helics::Federate>("test1", fi);
 
     gmlc::libguarded::guarded<std::vector<std::pair<int, std::string>>> mlog;
-    Fed->setLoggingCallback([&mlog](int level, const std::string&, const std::string& message) {
-        mlog.lock()->emplace_back(level, message);
-    });
+    Fed->setLoggingCallback(
+        [&mlog](int level, const std::string& /*unused*/, const std::string& message) {
+            mlog.lock()->emplace_back(level, message);
+        });
 
     Fed->enterExecutingMode();
     Fed->logMessage(3, "test MEXAGE1");
@@ -242,4 +256,39 @@ TEST(logging_tests, check_log_message_levels_high)
         }
     }
     EXPECT_TRUE(found_low && found_high);
+}
+
+TEST(logging_tests, dumplog)
+{
+    helics::FederateInfo fi(CORE_TYPE_TO_TEST);
+    fi.coreInitString = "--autobroker";
+    fi.setProperty(helics::defs::log_level, -1);
+
+    auto Fed = std::make_shared<helics::Federate>("test1", fi);
+    auto cr = Fed->getCorePointer();
+    gmlc::libguarded::guarded<std::vector<std::pair<int, std::string>>> mlog;
+    cr->setLoggingCallback(helics::local_core_id,
+                           [&mlog](int level,
+                                   const std::string& /*unused*/,
+                                   const std::string& message) {
+                               mlog.lock()->emplace_back(level, message);
+                           });
+
+    Fed->enterExecutingMode();
+    /** We are setting the flag then clearing it
+    this will generate 1 and at most 2 messages in the log callback
+    Thus the check for this is that there is a least 2 and at most 3 messages
+    in the log block, to indicate that the set and clear was successful*/
+    Fed->setFlagOption(helics_flag_dumplog);
+    Fed->setFlagOption(helics_flag_dumplog, false);
+
+    Fed->finalize();
+    cr->waitForDisconnect();
+    cr.reset();
+    auto llock = mlog.lock();
+    EXPECT_GE(llock->size(), 2U);
+    EXPECT_LE(llock->size(), 3U);
+    // this is to check that it has the correct level
+    EXPECT_EQ(llock->back().first, -10);  // the -10 should have a level enum value at some point in
+                                          // the future as part of the debugging improvements
 }
